localnews/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
flask-jwt-extended==4.4.4
flask-cors==4.0.0
Flask-Caching==2.0.1
celery==5.3.0
redis==4.5.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
pytest==7.4.3
feedparser==6.0.10
geopy==2.4.1
beautifulsoup4==4.12.3
